anuvaad-etl/anuvaad-extractor/file_translator/etl-file-translator/requirements.txt
pip==20.1
nltk==3.5
kafka-python==2.0.1
flask==1.1.2
flask-cors==3.0.8
flask-restful==0.3.8
elasticsearch==7.8.0
anuvaad-auditor==0.1.3
requests==2.25.1
python-docx==0.8.11
python-pptx==0.6.19
PyDocX==0.9.10
beautifulsoup4==4.9.3